requirements.txt
beautifulsoup4==4.13.4
firebase_admin==7.1.0
openai==1.97.1
playwright==1.54.0
python-dotenv==1.1.1
Requests==2.32.4
langchain
langchain-openai
pydantic
inflect